pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "business-platform-fastapi"
version = "1.0.0"
description = "FastAPI Backend for Business Platform with Admin Panel"
authors = [{name = "TKO Team"}]
license = "MIT"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "sqlalchemy>=2.0.0",
    "alembic>=1.12.0",
    "psycopg2-binary>=2.9.0",
    "asyncpg>=0.29.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "python-jose[cryptography]>=3.3.0",
    "PyJWT>=2.8.0",
    "passlib[bcrypt]>=1.7.4",
    "python-multipart>=0.0.6",
    "sqladmin>=0.16.1",
    # removed fastapi-admin/tortoise/aerich in favor of SQLAdmin with SQLAlchemy
    "redis>=5.0.0",
    "celery>=5.3.0",
    "httpx>=0.25.0",
    "python-dotenv>=1.0.0",
    "pillow>=10.0.0",
    "jinja2>=3.1.0",
    "aiofiles>=23.0.0",
    "email-validator>=2.1.0",
    "cryptography>=41.0.0",
    "imapclient>=2.3.1",
    "bcrypt>=4.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "httpx>=0.25.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "flake8>=6.0.0",
    "mypy>=1.7.0",
    "pre-commit>=3.5.0",
]

## CLI scripts were removed to satisfy build validation inside Docker. Use Docker CMD.

[tool.black]
line-length = 88
target-version = ['py39']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
